feat: cli add save-rendered flag (#822)
* feat: add save rendered flag to test command

* gix: move save rendered to executor for testing

* test: add flag test

* fix: linter issues

* fix: make save rendered bool

* ref: remove comment

+++ cmd/kustomize/main.go
@@ -71,7 +71,10 @@ func New(testCtx *test.TestCommandContext, kustomizeCtx *KustomizeContext) *cobr
 			if err != nil {
 				return err
 			}
-			defer os.Remove(tempFilename)
+
+			if !testCommandFlags.SaveRendered {
+				defer os.Remove(tempFilename)
+			}
 
 			err = test.TestWrapper(testCtx, []string{tempFilename}, testCommandFlags)
 			if err != nil {
